Use exactly same styling for tx list filter picker for iOS and Android
This required a styling hack to place a custom dropdown button on top of Android native picker.
