+++ src/components/dashboard/TxListHeader.js
@@ -1,7 +1,7 @@
 import { View, Text } from '../common'
 import ScanIndicator from './ScanIndicator'
 import PropTypes from 'prop-types'
-import Filter from './Filter'
+import Filter from './filter/Filter'
 import React from 'react'
 
 class TxListHeader extends React.Component {
@@ -22,7 +22,14 @@ class TxListHeader extends React.Component {
 
   render() {
     return (
-      <View justify="space-between" align="center" row bg="primary" py={1}>
+      <View
+        justify="space-between"
+        align="center"
+        row
+        bg="primary"
+        py={1}
+        pr={1}
+      >
         <View row align="center">
           <Text size="medium" pl={2} mr={1} py={1} shadow weight="semibold">
             Transactions
