+++ src/components/dashboard/filter/Filter.ios.js
@@ -1,7 +1,7 @@
-import { View, Text, BaseBtn } from '../common'
-import { Svg, G, Line } from 'react-native-svg'
+import { View, BaseBtn } from '../../common'
+import DropdownBtn from './DropdownBtn'
 import PropTypes from 'prop-types'
-import theme from '../../theme'
+import theme from '../../../theme'
 import React from 'react'
 import RN from 'react-native'
 
@@ -23,46 +23,13 @@ class Filter extends React.Component {
   }
 
   render() {
-    const shouldCondenseLabel =
-      ['received', 'converted', 'auction'].includes(this.props.filter) &&
-      RN.Dimensions.get('window').width < 375
-
     return (
       <View pr={1}>
         <RN.TouchableOpacity activeOpacity={0.9} onPress={this.toggleDropdown}>
-          <View
-            style={styles.dropdownBtn}
-            align="center"
-            row
-            bg={this.state.isOpen ? 'dark' : 'transparent'}
-            py={1.25}
-            px={1}
-          >
-            <Text
-              opacity={1}
-              weight="semibold"
-              shadow
-              size={shouldCondenseLabel ? 'xSmall' : 'small'}
-              ls={shouldCondenseLabel ? 0 : 1.4}
-              mr={0.5}
-            >
-              {this.props.options[this.props.filter]}
-            </Text>
-            <Svg viewBox="0 2 21 20" width="14" height="12">
-              <G
-                strokeLinecap="round"
-                strokeWidth="2"
-                fillRule="evenodd"
-                stroke="white"
-                fill="none"
-              >
-                <G rotation={this.state.isOpen ? '90' : '-90'} origin="10, 13">
-                  <Line x1="7" y1="13" x2="15" y2="5" />
-                  <Line x1="7" y1="14" x2="15" y2="22" />
-                </G>
-              </G>
-            </Svg>
-          </View>
+          <DropdownBtn
+            label={this.props.options[this.props.filter]}
+            isOpen={this.state.isOpen}
+          />
         </RN.TouchableOpacity>
         <View style={styles.dropdownContainer}>
           {this.state.isOpen && (
@@ -91,10 +58,6 @@ class Filter extends React.Component {
 const styles = RN.StyleSheet.create({
   dropdownContainer: {
     position: 'relative'
-  },
-  dropdownBtn: {
-    borderTopLeftRadius: 4,
-    borderTopRightRadius: 4
   },
   dropdown: {
     backgroundColor: theme.colors.dark,
